+++ src/main/java/org/example/coworking/infrastructure/dao/ModelDao.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface ModelDao<T, E extends Exception> {
     void add(T object);
     void delete(T object) throws E;
-    T getById(int id) throws E;
+    T getById(Long id) throws E;
     List<T> getAll();
     void load();
     void save();
